+++ libc8/c8/graphics.h
@@ -11,6 +11,8 @@
 #ifndef C8_GRAPHICS_H
 #define C8_GRAPHICS_H
 
+#include <stdint.h>
+
 #define C8_LOW_DISPLAY_WIDTH 64
 #define C8_LOW_DISPLAY_HEIGHT 32
 #define C8_HIGH_DISPLAY_WIDTH 128
@@ -33,16 +35,16 @@
  * @param y y offset (for `DISPLAY_EXTENDED`)
  */
 typedef struct {
-	int p[C8_HIGH_DISPLAY_WIDTH * C8_HIGH_DISPLAY_HEIGHT];
-	int mode;
-	int x, y;
+	uint8_t p[C8_HIGH_DISPLAY_WIDTH * C8_HIGH_DISPLAY_HEIGHT];
+	uint8_t mode;
+	uint8_t x, y;
 } c8_display_t;
 
-int *c8_get_pixel(c8_display_t *, int, int);
+uint8_t *c8_get_pixel(c8_display_t *, int, int);
 
 extern void c8_beep(void);
 extern void c8_deinit_graphics(void);
-extern int c8_init_graphics(void);
+extern uint8_t c8_init_graphics(void);
 extern void c8_render(c8_display_t *, int *);
 extern int c8_tick(int *, int);
